@@ -11,15 +11,19 @@
     "start:dev": "sirv public --single --dev"
   },
   "devDependencies": {
+    "@budibase/client": "*",
+    "lodash": "^4.17.15",
     "npm-run-all": "^4.1.5",
     "rollup": "^1.11.0",
     "rollup-plugin-commonjs": "^10.0.2",
+    "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-livereload": "^1.0.1",
     "rollup-plugin-node-resolve": "^5.0.0",
     "rollup-plugin-svelte": "^5.0.0",
     "rollup-plugin-terser": "^5.1.1",
+    "shortid": "^2.2.15",
     "sirv-cli": "^0.4.4",
-    "svelte": "^3.0.0"
+    "svelte": "^3.12.1"
   },
   "keywords": [
     "svelte"
