@@ -5,6 +5,8 @@ import Textbox from "../common/Textbox.svelte";
 import Dropdown from "../common/Dropdown.svelte";
 import ComponentPropSelector from "./ComponentPropSelector.svelte";
 import PropArraySelector from "./PropArraySelector.svelte";
+import EventListSelector from "./EventListSelector.svelte";
+import StateBindingControl from "./StateBindingControl.svelte";
 
 export let errors = [];
 export let setProp = () => {};
@@ -26,37 +28,39 @@ const setComponentProp = (props) => {
 
 <div class="root" >
 
+    {#if propDef.type === "component"}
 
-    {#if propDef.type === "bool"}
-    <Checkbox label={propDef.____name} 
-                checked={props[propDef.____name]} 
-                on:change={ev => setProp(propDef.____name, ev.target.checked)}
-                hasError={fieldHasError(propDef.____name)} />
-    {:else if propDef.type === "options"}
-    <Dropdown label={propDef.____name}
-                selected={props[propDef.____name]} 
-                options={propDef.options}
-                on:change={ev => setProp(propDef.____name, ev.target.value)}
-                hasError={fieldHasError(propDef.____name)}/>
-    {:else if propDef.type === "component"}
+    <div class="prop-label">{propDef.____name}</div>
     <ComponentPropSelector label={propDef.____name}
                             props={props[propDef.____name]}
                             {disabled}
                             onEdit={onEditComponent}
                             onComponentChosen={onEditComponent}
                             onValueChanged={setComponentProp}/>
+
     {:else if propDef.type === "array"}
+
+    <div class="prop-label">{propDef.____name}</div>
     <PropArraySelector parentProps={props}
                        {propDef}
                        onValueChanged={setComponentProp}
-                       onEditComponentProp={onEditComponent}/>
+                       onEditComponentProp={onEditComponent} />
+
+    {:else if propDef.type === "event"}
+
+    <div class="prop-label">{propDef.____name}</div>
+    <EventListSelector parentProps={props}
+                       {propDef}
+                       onValueChanged={setComponentProp} />
+
     {:else}
-    <Textbox label={propDef.____name}
-                text={props[propDef.____name]}
-                on:change={ev => setProp(propDef.____name, ev.target.value)}
-                margin={false}
-                hasError={fieldHasError(propDef.____name)}
-                {disabled}/>
+
+    <div class="prop-label">{propDef.____name}</div>
+    <StateBindingControl value={props[propDef.____name]}
+                         type={propDef.type}
+                         options={propDef.options}
+                         onChanged={v => setProp(propDef.____name, v)}/>
+
     {/if} 
 
 </div>
@@ -68,6 +72,12 @@ const setComponentProp = (props) => {
     border-style: dotted;
     border-width: 0 0 1px 0;
     border-color: var(--primary25);
+}
+
+.prop-label {
+    font-size: 0.9em;
+    color: var(--primary100);
+    font-weight: bold;
 }
 
 </style>
