@@ -100,16 +100,19 @@ const getPermissionsString = perms => {
 <style>
 
 .root {
-    padding:10px;
+    height: 100%;
+    position: relative;
+    padding: 1.5rem;
+
 }
 
-.edit-button {
-    cursor:pointer;
-    color: var(--white);
+.actions-header {
+    flex: 0 1 auto;
 }
 
-tr:hover .edit-button  {
-    color: var(--secondary75);
+.node-view {
+    overflow-y: auto;
+    flex: 1 1 auto;
 }
 
 
