@@ -1,3 +1,3 @@
 import feather from "feather-icons";
-const getIcon = (icon, size) => feather.icons[icon].toSvg({height:size||"24", width:size||"24"});
+const getIcon = (icon, size) => feather.icons[icon].toSvg({height:size||"16", width:size||"16"});
 export default getIcon;
