@@ -10,7 +10,7 @@ import {
 } from "lodash/fp";
 import { fade, slide } from 'svelte/transition';
 
-export let propertyName = "";
+export let title = "";
 export let onGoBack = () => {};
 export let instanceProps = {};
 export let onPropsChanged = () => {};
@@ -18,6 +18,9 @@ export let onPropsChanged = () => {};
 
 let editingSubComponentName;
 let editingSubComponentProps;
+let editingSubComponentArrayIndex;
+let editingSubComponentArrayPropName;
+let editingSubComponentTitle;
 let allComponents;
 
 store.subscribe(s => {
@@ -33,18 +36,28 @@ const onSubComponentGoBack = () => {
 }
 
 const onEditComponentProp = (propName, arrayIndex, arrayPropName) => {
-    editingSubComponentName = isUndefined(arrayIndex)
+    editingSubComponentName = propName;
+    editingSubComponentTitle = isUndefined(arrayIndex)
                               ? propName
                               : `${propName}[${arrayIndex}].${arrayPropName}`;
     editingSubComponentProps = isUndefined(arrayIndex)
                                ? instanceProps[propName]
                                : instanceProps[propName][arrayIndex][arrayPropName];
+    editingSubComponentArrayIndex = arrayIndex;
+    editingSubComponentArrayPropName = arrayPropName;
 };
 
 
 const onSubComponentPropsChanged = (subProps) => {
     const newProps = cloneDeep(instanceProps);
-    newProps[editingSubComponentName] = subProps;
+    if(isUndefined(editingSubComponentArrayIndex)) {
+        newProps[editingSubComponentName] = subProps;
+    } else {
+        newProps[editingSubComponentName]
+                [editingSubComponentArrayIndex]
+                [editingSubComponentArrayPropName] = subProps;
+    }
+    
     instanceProps = newProps;
     onPropsChanged(newProps);
 }
@@ -62,7 +75,7 @@ const propsChanged = newProps => {
     <div class="title">
         <IconButton icon="chevron-left"
                     on:click={onGoBack}/>
-        <span>{propertyName}</span>
+        <span>{title}</span>
     </div>
 
     {#if editingSubComponentName}
@@ -71,7 +84,7 @@ const propsChanged = newProps => {
         <svelte:self onPropsChanged={onSubComponentPropsChanged}
                     onGoBack={onSubComponentGoBack}
                     instanceProps={editingSubComponentProps}
-                    propertyName={editingSubComponentName} />
+                    title={editingSubComponentTitle} />
     </div>
     {:else}
     <PropsView {instanceProps}
