@@ -3,8 +3,14 @@ import {
     isBoolean, 
     isNumber,
     isArray,
-    isObjectLike
+    isObjectLike,
+    isPlainObject,
+    every
 } from "lodash/fp";
+
+import {
+    EVENT_TYPE_MEMBER_NAME
+} from "../../common/eventHandlers";
 
 const defaultDef = typeName => () => ({
     type: typeName,
@@ -47,6 +53,14 @@ export const expandPropsDefinition = propsDefinition => {
 }
 
 const isComponent = isObjectLike;
+const isEvent = e => 
+    isPlainObject(e) 
+    && isString(e[EVENT_TYPE_MEMBER_NAME])
+    && isPlainObject(e.parameters);
+
+const isEventList = e => 
+    isArray(e) && every(isEvent)(e);
+
 
 export const types = {
     string: propType(() => "", isString, defaultDef("string")),
@@ -56,4 +70,5 @@ export const types = {
     options: propType(() => "", isString, defaultDef("options")),
     component: propType(() => ({_component:""}), isComponent, defaultDef("component")),
     asset: propType(() => "", isString, defaultDef("asset")),
+    event: propType(() => [], isEventList, defaultDef("event"))
 };
