@@ -4,7 +4,12 @@ import resolve from 'rollup-plugin-node-resolve';
 export default {
 	input: 'src/index.js',
 	output: [
-		{ file: "dist/index.js", 'format': 'esm', name:"budibaseStandardComponents" }
+		{ 
+			file: "dist/index.js", 
+			format: 'esm', 
+			name:"budibaseStandardComponents",
+			sourcemap: "inline"
+		}
 	],
 	plugins: [
 		svelte(),
