@@ -3,8 +3,24 @@ import resolve from 'rollup-plugin-node-resolve';
 import commonjs from 'rollup-plugin-commonjs';
 import livereload from 'rollup-plugin-livereload';
 import { terser } from 'rollup-plugin-terser';
+import json from 'rollup-plugin-json';
 
 const production = !process.env.ROLLUP_WATCH;
+
+const lodash_fp_exports = ["union", "reduce", "isUndefined", "cloneDeep", "split", "some", "map", "filter", "isEmpty", "countBy", "includes", "last", "find", "constant", 
+"take", "first", "intersection", "mapValues", "isNull", "has", "isNumber", "isString", "isBoolean", "isDate", "isArray", "isObject", "clone", "values", "keyBy", 
+"keys", "orderBy", "concat", "reverse", "difference", "merge", "flatten", "each", "pull", "join", "defaultCase", "uniqBy", "every", "uniqWith", "isFunction", "groupBy", 
+"differenceBy", "intersectionBy", "isEqual", "max", "sortBy", "assign", "uniq", "trimChars", "trimCharsStart", "isObjectLike", "flattenDeep", "indexOf"];
+
+const lodash_exports = ["toNumber", "flow", "isArray", "join", "replace", "trim", "dropRight", "takeRight", "head", "isUndefined", "isNull", "isNaN", "reduce", "isEmpty", 
+"constant", "tail", "includes", "startsWith", "findIndex", "isInteger", "isDate", "isString", "split", "clone", "keys", "isFunction", "merge", "has", "isBoolean", "isNumber", 
+"isObjectLike", "assign", "some", "each", "find", "orderBy", "union", "cloneDeep"];
+
+const coreExternal = [
+	"lodash", "lodash/fp", "date-fns",
+	"lunr", "safe-buffer", "shortid",
+	"@nx-js/compiler-util"
+];
 
 export default {
 	input: 'src/Test/testMain.js',
@@ -32,9 +48,20 @@ export default {
 		// https://github.com/rollup/rollup-plugin-commonjs
 		resolve({
 			browser: true,
-			dedupe: importee => importee === 'svelte' || importee.startsWith('svelte/')
+			dedupe: importee => {
+				return importee === 'svelte' 
+					   || importee.startsWith('svelte/')
+					   || coreExternal.includes(importee);
+			}
 		}),
-		commonjs(),
+		commonjs({
+			namedExports: {
+				"lodash/fp": lodash_fp_exports,
+				"lodash":lodash_exports,
+				"shortid": ["generate"]
+			}
+		}),
+		json(),
 
 		// Watch the `public` directory and refresh the
 		// browser on changes when not in production
