@@ -1,18 +1,34 @@
 <script>
 
 import Textbox from "./Textbox.svelte";
-import FormControl from "./FormControl.svelte";
 import Form from "./Form.svelte";
 import Button from "./Button.svelte";
+import { authenticate } from "./api";
 
 export let usernameLabel = "Username";
 export let passwordLabel = "Password";
 export let loginButtonLabel = "Login";
 export let loginRedirect = "";
-export let logo = "/budibase-logo.png";
+export let logo = "";
+export let buttonClass = "";
 
 let username = "";
 let password = "";
+let busy = false;
+let incorrect = false;
+
+const login = () => {
+    busy = true;
+    authenticate(username, password)
+    .then(r => {
+        busy = false;
+        if(r.status === 200) {
+            // reload page
+        } else {
+            incorrect = true;
+        }
+    })
+}
 
 </script>
 
@@ -20,22 +36,40 @@ let password = "";
 
     <div class="content">
 
+        {#if logo}
         <div class="logo-container">
             <img src={logo} alt="logo"/>
         </div>
+        {/if}
 
-        <Form>
-            <FormControl label={usernameLabel}>
+        <div class="form-root">
+            <div class="label">
+                {usernameLabel}
+            </div>
+            <div class="control">
                 <Textbox bind:value={username} />
-            </FormControl>
-            <FormControl label={passwordLabel}>
+            </div>
+            <div class="label">
+                {passwordLabel}
+            </div>
+            <div class="control">
                 <Textbox bind:value={password} hideValue=true />
-            </FormControl>
-        </Form>
+            </div>
+        </div>
 
         <div class="login-button-container">
-            <Button>{loginButtonLabel}</Button>
+            <Button disabled={busy} 
+                    on:click={login}
+                    class={buttonClass}>
+                    {loginButtonLabel}
+            </Button>
         </div>
+
+        {#if incorrect}
+        <div class="incorrect-details-panel">
+            Incorrect username or password
+        </div>
+        {/if}
 
     </div>
 
@@ -67,6 +101,33 @@ let password = "";
 .login-button-container {
     text-align: right;
     margin-top: 20px;
+}
+
+.incorrect-details-panel {
+    margin-top: 30px;
+    padding: 10px;
+    border-style: solid;
+    border-width: 1px;
+    border-color: maroon;
+    border-radius: 1px;
+    text-align: center;
+    color: maroon;
+    background-color: mistyrose;
+}
+
+.form-root {
+    display: grid;
+    grid-template-columns: [label] auto [control] 1fr; /* [overflow] auto;*/
+}
+
+.label {
+    grid-column-start: label;
+    padding: 5px 10px;
+    vertical-align: middle;
+}
+.control {
+    grid-column-start: control;
+    padding: 5px 10px;
 }
 
 </style>
