@@ -1,6 +1,6 @@
 export {default as button} from  "./Button.svelte";
 export {default as login} from  "./Login.svelte";
 export {default as form} from  "./Form.svelte";
-export {default as formControl} from  "./FormControl.svelte";
+export {default as grid} from  "./Grid.svelte";
 export {default as textbox} from  "./Textbox.svelte";
 export {default as stackpanel} from  "./StackPanel.svelte";
